adding tags, moving history controller
